requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.1.13
langchain-groq==0.1.3
python-dotenv==1.0.1
pydantic==2.6.4